Fixed File::Copy::Requires usage for Windows.
File::Copy::Requires is only needed during the Windows build stage,
but I was "use"ing it at the top of the build module, which means
it was being requested at the configure stage, when it wasn't
necessarily available. I've moved the "use" to a "require" within
the function that actually needs it.